# openapi.yaml
openapi: "3.1.1"
info:
  title: Rave Collection E-commerce API
  description: API for Rave Collection shoe shop e-commerce and inventory management system
  version: 1.0.0
  contact:
    name: Rave Collection
    email: support@ravecollection.com
servers:
  - url: http://localhost:5005/api
    description: Development server
security:
  - bearerAuth: []
paths:
...
  /products:
    get:
      tags:
        - Products
      summary: Get all products
      security: []
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      tags:
        - Products
      summary: Create new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      tags:
        - Products
      summary: Get product by ID
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      tags:
        - Products
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      tags:
        - Products
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Product deleted
  /products/{id}/stock:
    put:
      tags:
        - Products
      summary: Update product stock
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - stock
              properties:
                stock:
                  type: integer
      responses:
        "200":
          description: Stock updated

  # Category endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: decimal
        brand:
          type: string
        category_id:
          type: integer
        stock:
          type: integer
        image_url:
          type: string
        created_at:
          type: string
          format: date-time
    ProductInput:
      type: object
      required:
        - name
        - price
        - brand
        - category_id
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: decimal
        brand:
          type: string
        category_id:
          type: integer
        stock:
          type: integer
          default: 0
        image_url:
          type: string
...